@@ -32,32 +32,31 @@
   "typings": "index.d.ts",
   "homepage": "https://github.com/Greentube/localize-router",
   "peerDependencies": {
-    "@angular/common": ">=6.0.0",
-    "@angular/core": ">=6.0.0",
-    "@angular/router": ">=6.0.0",
-    "@ngx-translate/core": ">=10.0.0",
-    "rxjs": "^6.0.0"
+    "@angular/common": ">=9.0.5",
+    "@angular/core": ">=9.0.5",
+    "@angular/router": ">=9.0.5",
+    "@ngx-translate/core": ">=12.0.0",
+    "rxjs": "^6.5.3"
   },
   "devDependencies": {
-    "@angular/animations": "~6.0.6",
-    "@angular/common": "~6.0.6",
-    "@angular/compiler": "~6.0.6",
-    "@angular/compiler-cli": "~6.0.6",
-    "@angular/core": "~6.0.6",
-    "@angular/http": "~6.0.6",
-    "@angular/platform-browser": "~6.0.6",
-    "@angular/platform-browser-dynamic": "~6.0.6",
-    "@angular/platform-server": "~6.0.6",
-    "@angular/router": "~6.0.6",
-    "@ngx-translate/core": "10.0.2",
+    "@angular/animations": "~9.0.5",
+    "@angular/common": "~9.0.5",
+    "@angular/compiler": "~9.0.5",
+    "@angular/compiler-cli": "~9.0.5",
+    "@angular/core": "~9.0.5",
+    "@angular/platform-browser": "~9.0.5",
+    "@angular/platform-browser-dynamic": "~9.0.5",
+    "@angular/platform-server": "~9.0.5",
+    "@angular/router": "~9.0.5",
+    "@ngx-translate/core": "^12.0.0",
     "@types/es6-promise": "0.0.32",
     "@types/hammerjs": "2.0.34",
     "@types/jasmine": "2.5.47",
-    "@types/node": "7.0.22",
+    "@types/node": "~13.9.0",
     "awesome-typescript-loader": "3.1.3",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "~5.2.1",
     "concurrently": "3.1.0",
-    "core-js": "2.4.1",
+    "core-js": "~3.6.4",
     "istanbul-instrumenter-loader": "~2.0.0",
     "jasmine-core": "2.6.2",
     "karma": "1.7.0",
@@ -70,16 +69,17 @@
     "karma-webpack": "2.0.3",
     "loader-utils": "1.1.0",
     "reflect-metadata": "0.1.10",
-    "rxjs": "6.2.1",
+    "rxjs": "6.5.3",
     "source-map-loader": "0.2.1",
     "ts-helpers": "1.1.2",
+    "tslib": "~1.10.0",
     "tslint": "5.8.0",
     "tslint-eslint-rules": "^4.1.1",
     "tslint-loader": "3.5.3",
-    "typescript": "2.7.2",
+    "typescript": "3.7.5",
     "webpack": "2.6.0",
     "webpack-node-externals": "^1.5.4",
     "yarn": "^1.12.1",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.10.2"
   }
 }
