@@ -1,8 +1,7 @@
 
 Error.stackTraceLimit = Infinity;
 
-require('core-js/es6');
-require('core-js/es7/reflect');
+require('core-js');
 
 // Typescript emit helpers polyfill
 require('ts-helpers');
